pyproject.toml
[tool.poetry]
name = "ska-src-maltopuft-backend"
version = "0.0.1"
description = "MALTOPUFT backend service"
authors = ["Elliot Goodwin <elliot.goodwin@manchester.ac.uk>"]
readme = ["README.md", "LICENSE"]
license = "BSD-3-Clause"
packages = [
    { include = "ska_src_maltopuft_backend", from = "src" }
]


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"


[[tool.poetry.source]]
name = "ska-nexus"
url = "https://artefact.skao.int/repository/pypi-internal/simple"
priority = "supplemental"


[tool.poetry.dependencies]
python = ">=3.10,<4.0.0"
fastapi = "^0.110.2"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
ska-ser-logging = "^0.4.3"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.29"
psycopg = {version = "^3.1.18", extras = ["binary", "pool", "c"]}
httpx = "^0.27.0"
itsdangerous = "^2.2.0"
requests = "^2.31.0"
pyjwt = "^2.8.0"
alembic = "^1.13.1"
email-validator = "^2.1.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
pytest = "^8.0.0"
pytest-cov = "^3.0"
pytest-json-report = "^1.5.0"
pylint = "^3.1.0"
black = "^24.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
pylint-junit = "^0.3.4"
ruff = "^0.3.4"
pytest-bdd = "^7.1.2"
pytest-mock = "^3.14.0"
faker = "^25.8.0"
pytest-asyncio = "^0.23.7"
polyfactory = "^2.16.0"


[tool.poetry.group.docs]
optional = true
[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"
recommonmark = "^0.7.1"
ska-ser-sphinx-theme = "^0.1.2"
myst-parser = "^2.0.0"
sphinx-autoapi = "^3.0.0"


[tool.pytest.ini_options]
minversion = "6.0"
addopts = """
    --cov src
    --cov-report=term-missing
    --cov-report html:build/reports/code-coverage
"""


testpaths = ["tests"]

[tool.coverage.run]
source = ["./src"]
omit = ["**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "**/__init__.py",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "docs",
    "node_modules",
    "site-packages",
    "test",
    "venv",
    ".make",
]

# Assume Python 3.11
target-version = "py311"


[tool.ruff.lint]
# Pylint, flake8 and isort replacement
select = ["ALL"]
ignore = [
    "ANN101",  # missing-type-self
    "ANN401",  # any-type
    "B008",  # function-call-in-default-argument
    "D203", "D213",  # Silence incompatability warnings
    "D205",  # blank-line-after-summary
    "D400",  # ends-in-period
    "D401",  # non-imperative-mood
    "D407",  # dashed-underline-after-section
    "D415",  # ends-in-punctuation
    "G004",  # logging-f-string
    "INP001",  # implicit-namespace-package
    "PT012",  # `pytest.raises()` block should contain a single simple statement
    "S101", "S104",  # assert, possible binding to all interfaces
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.mypy]
plugins = ["pydantic.mypy"]
exclude = [
    "./docs",
    "./tests",
]

[[tool.mypy.overrides]]
module = [
    "ska_ser_logging",
]
ignore_missing_imports = true


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
